5.18-导出接口到ios。取消接口日志，防止再IOS上不能创建日志产生错误。修改接口声明和实现参数不一致导致的Bug。优化代码结构。

+++ TCPServer3.0/MyServer.cpp
@@ -139,7 +139,7 @@ void MyServer::OnNetMsgRecv(XClient* pClient, MsgHeader* pMsgHeader)
 			s.WriteInt64(40);
 			s.WriteFloat(55.66f);
 			s.WriteDouble(77.88);
-			s.WriteArray("Server", strlen("Server"));
+			s.WriteArray("Server", (int)strlen("Server"));
 			s.Finish(MSG_BYTESTREAM);
 			
 			if (pClient->SendStream(&s) < 0)
